chore:  cleanup in Class class
