+++ src/types/Character/Class/helpers.ts
@@ -1,6 +1,22 @@
 import { Skill } from '../Skills';
 import { Aura, AurasList } from './Aura';
 import { ClassFeature, FeaturesList } from './Features';
+import type { Class } from './index';
+
+interface SavesProgression {
+    fortitude: boolean;
+    reflexes: boolean;
+    will: boolean;
+}
+
+interface CastingProgression {
+    arcane: number;
+    divine: number;
+    spontaneous: number;
+}
+
+type Classes = { [key: string]: Class };
+type ClassLevels = { cls: Class; level: number }[];
 
 const validSkills = new Set(Object.values(Skill));
 
@@ -36,3 +52,5 @@ function filterSkills(raw: string[]): Set<Skill> {
 }
 
 export { validSkills, hdAverage, hdFirst, mapFeatures, mapAuras, filterSkills };
+export type { Classes, ClassLevels, SavesProgression, CastingProgression };
+
